Changes to implement a renderer for scanned items and fix to create a PDF from the scanned items
