+++ WeScan/MultiPageSession/MultiPageScanSession.swift
@@ -11,41 +11,6 @@ import Foundation
 public struct ScannedItem:Equatable{
     let originalImage:UIImage
     var quad:Quadrilateral?
-    
-    public func renderQuadImage(completion: @escaping (_ image:UIImage?)->Void){
-        DispatchQueue.global(qos: .background).async {
-            let image = self.originalImage.applyingPortraitOrientation()
-            
-            guard let quad = self.quad,
-                let ciImage = CIImage(image: image) else {
-                    // TODO: Return error
-                    DispatchQueue.main.async { completion(nil) }
-                    return
-            }
-            
-            var cartesianScaledQuad = quad.toCartesian(withHeight: image.size.height)
-            cartesianScaledQuad.reorganize()
-            
-            let filteredImage = ciImage.applyingFilter("CIPerspectiveCorrection", parameters: [
-                "inputTopLeft": CIVector(cgPoint: cartesianScaledQuad.bottomLeft),
-                "inputTopRight": CIVector(cgPoint: cartesianScaledQuad.bottomRight),
-                "inputBottomLeft": CIVector(cgPoint: cartesianScaledQuad.topLeft),
-                "inputBottomRight": CIVector(cgPoint: cartesianScaledQuad.topRight)
-                ])
-            
-            //let enhancedImage = filteredImage.applyingAdaptiveThreshold()?.withFixedOrientation()
-            
-            var uiImage: UIImage!
-            
-            // Let's try to generate the CGImage from the CIImage before creating a UIImage.
-            if let cgImage = CIContext(options: nil).createCGImage(filteredImage, from: filteredImage.extent) {
-                uiImage = UIImage(cgImage: cgImage)
-            } else {
-                uiImage = UIImage(ciImage: filteredImage, scale: 1.0, orientation: .up)
-            }
-            DispatchQueue.main.async { completion(uiImage.withFixedOrientation()) }
-        }
-    }
 }
 
 public class MultiPageScanSession {
