+++ WeScanSampleProject/HomeViewController.swift
@@ -84,12 +84,7 @@ final class HomeViewController: UIViewController {
         let scannerVC = ImageScannerController()
         scannerVC.imageScannerDelegate = self
         present(scannerVC, animated: true, completion: nil)
-        
-        /*
-        let imageScanner = ImageScanner()
-        imageScanner.startScanningFrom(baseViewController: self)*/
     }
-    
 }
 
 extension HomeViewController: ImageScannerControllerDelegate {
